fix: improve error logging on mutation

+++ backend/main/mutations/UserMutations.py
@@ -1,4 +1,5 @@
-from graphene import String, Mutation, Field, ID, Boolean, ResolveInfo
+from graphene import String, Mutation, Field, ID, Boolean, ResolveInfo, List
+from graphene_django.types import ErrorType
 
 from main.types.UserType import UserType
 from django.contrib.auth import get_user_model
@@ -7,20 +8,21 @@ class CreateUser(Mutation):
 
     user = Field(UserType)
 
+    class Arguments:
+        username = String(required=True)
         email = String(required=True)
-
-    user = Field(UserType)
     
     @classmethod
     def mutate(cls, root: None, info: ResolveInfo, username: str, email: str):
         user = get_user_model()(username=username, email=email)
         user.save()
-        return CreateUser(user=user)
+        return cls(user=user)
 
 
 class UpdateUser(Mutation):
 
     user = Field(UserType)
+    errors = List(ErrorType)
 
     class Arguments:
         id = ID(required=True)
@@ -29,10 +31,15 @@ class UpdateUser(Mutation):
 
     @classmethod
     def mutate(cls, root: None, info: ResolveInfo, username: str, email: str):
-        try:
-            user = get_user_model().objects.get(pk=id)
-        except get_user_model().DoesNotExist:
-            raise Exception("user not found")
+
+        user = get_user_model().objects.get(pk=id)
+        
+        if user is None:
+            error = ErrorType(
+                field="id",
+                messages=["User not found."],
+            )
+            return cls(errors=[error])
 
         if username is not None:
             user.username = username
@@ -40,22 +47,27 @@ class UpdateUser(Mutation):
             user.email = email
 
         user.save()
-        return UpdateUser(user=user)
+        return cls(user=user)
 
 
 class DeleteUser(Mutation):
 
-    success = Boolean()
-    
+    ok = Boolean()
+    errors = List(ErrorType)
+
     class Arguments:
         id = ID(required=True)
 
     @classmethod
     def mutate(cls, root: None, info: ResolveInfo, username: str, email: str):
-        try:
-            user = get_user_model().objects.get(pk=id)
-        except get_user_model().DoesNotExist:
-            raise Exception("User not found")
+        user = get_user_model().objects.get(pk=id)
+        
+        if user is None:
+            error = ErrorType(
+                field="id",
+                messages=["User not found."],
+            )
+            return cls(ok=False, errors=[error])
 
         user.delete()
-        return DeleteUser(success=True)
+        return cls(ok=True)
